Add notice to Citizens patch and remove ProtocolLib patch
