--- NachoSpigot-Server/src/main/java/me/elier/util/Utils.java
@@ -1,7 +1,0 @@
-package me.elier.util;
-
-public class Utils {
-    public static String repeat(String string, int count) {
-        return new String(new char[count]).replace("\0", string);
-    }
-}
