+++ NachoSpigot-Server/src/main/java/me/elier/util/minecraft/PluginUtils.java
@@ -1,0 +1,11 @@
+package me.elier.util.minecraft;
+
+import org.bukkit.plugin.Plugin;
+
+public class PluginUtils {
+
+    public static int getCitizensBuild(Plugin plugin) {
+       return Integer.parseInt(plugin.getDescription().getVersion().split("\\(build")[1].trim().replace(")", ""));
+    }
+
+}
